requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements.in'
#
annotated-types==0.6.0
    # via pydantic
anyio==4.0.0
    # via httpcore
blinker==1.6.2
    # via flask
build==0.10.0
    # via pip-tools
cachetools==5.3.1
    # via google-auth
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
click==8.1.5
    # via
    #   flask
    #   functions-framework
    #   pip-tools
cloudevents==1.9.0
    # via functions-framework
colorama==0.4.6
    # via
    #   build
    #   click
deprecation==2.1.0
    # via cloudevents
flask==2.3.2
    # via functions-framework
functions-framework==3.4.0
    # via -r .\requirements.in
google-api-core[grpc]==2.12.0
    # via google-cloud-secret-manager
google-auth==2.23.2
    # via google-api-core
google-cloud-secret-manager==2.16.4
    # via -r .\requirements.in
googleapis-common-protos[grpc]==1.60.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.12.6
    # via google-cloud-secret-manager
grpcio==1.59.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.59.0
    # via google-api-core
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
packaging==23.1
    # via
    #   build
    #   deprecation
pip-tools==6.13.0
    # via -r .\requirements.in
proto-plus==1.22.3
    # via google-cloud-secret-manager
protobuf==4.24.4
    # via
    #   google-api-core
    #   google-cloud-secret-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg2==2.9.9
    # via -r .\requirements.in
pyasn1==0.5.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.4.1
    # via -r .\requirements.in
pydantic-core==2.10.1
    # via pydantic
pyproject-hooks==1.0.0
    # via build
python-telegram-bot==20.5
    # via -r .\requirements.in
requests==2.31.0
    # via
    #   -r .\requirements.in
    #   google-api-core
rsa==4.9
    # via google-auth
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.0.3
    # via requests
watchdog==3.0.0
    # via functions-framework
werkzeug==2.3.6
    # via flask
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools

shared~=0.0.29